#2175 Changes to make deployer work

+++ app/Console/Commands/Environment/UpdatePrepare.php
@@ -3,7 +3,6 @@
 namespace App\Console\Commands\Environment;
 
 use App\Console\Commands\Traits\ExecutesShellCommands;
-use Artisan;
 use Illuminate\Console\Command;
 
 class UpdatePrepare extends Command
@@ -20,28 +19,30 @@ class UpdatePrepare extends Command
     /**
      * @var string
      */
-    protected $signature = 'environment:updateprepare {environment}';
+    protected $signature = 'environment:updateprepare';
 
     /**
      * Execute the console command.
      *
      * @return int
      */
-    public function handle()
+    public function handle(): int
     {
-        $environment = $this->argument('environment');
+        $environment = config('app_type');
 
-        $this->shell([
-            // Git commands
-            'git checkout .',
-            'git clean -f',
-            'git pull',
-        ]);
+        if ($environment === 'local') {
+            $this->shell([
+                // Git commands
+                'git checkout .',
+                'git clean -f',
+                'git pull',
+            ]);
+        }
 
         $this->shell([
             'npm install',
             'npm audit fix',
-            'node node_modules/puppeteer/install.js'
+            'node node_modules/puppeteer/install.js',
         ]);
 
         // Install composer here - a next command can then have the updated definitions of the autoloader when called
